refactored and documented a bit more
