+++ lib/log.js
@@ -1,0 +1,5 @@
+module.exports = function log(...messages) {
+  if (process.env.DEBUG) {
+    console.log(...messages)
+  }
+}
