+++ test/integration.test.js
@@ -4,7 +4,6 @@ const supertest = require('supertest')
 const express = require('express')
 const path = require('path')
 const bodyParser = require('body-parser')
-const jsdom = require('jsdom')
 
 describe("Page", () => {
